% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ts_country.R
\name{get_area_country}
\alias{get_area_country}
\title{Calculate the country coverage per grid cell}
\usage{
get_area_country(grid, country = dragon_country(), th = 0.5)
}
\arguments{
\item{grid}{a \code{terra::SpatVector} object with the grid}

\item{country}{a \code{terra::SpatVector} object with the country definition}

\item{th}{coverage threshold value to keep countries in}
}
\value{
A \code{data.frame} with the grid_id in rows and country in columns
}
\description{
Calculate the country coverage per grid cell
}
