% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bioclim_curve.R
\name{get_bioclim_curve}
\alias{get_bioclim_curve}
\title{Get bioclim curve}
\usage{
get_bioclim_curve(bioclim_seq, psi_coef, scaling, digits = 5)
}
\arguments{
\item{bioclim_seq}{List of bioclimatic variables values (names correspond to bioclimatic variables)}

\item{psi_coef}{psi coefficients file (original format)}

\item{scaling}{Scaling data (must have values corresponding to bioclimatic variables in the column variable)}

\item{digits}{integer indicating the number of decimal places to be kept.}
}
\value{
A dataframe with columns median, qmin, qmax, x_scaled, var (name of the bioclimatic variable) and species
}
\description{
Get values computed from bioclimatic coefficients that make the curve
}
