% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bioclim_seq.R
\name{get_bioclim_seq}
\alias{get_bioclim_seq}
\title{Get bioclim seq}
\usage{
get_bioclim_seq(env, length.out = 100)
}
\arguments{
\item{env}{Environment data (must have bioclimatic columns as bioxx_... columns)}

\item{length.out}{Length of the sequence of values to get}
}
\value{
List of sequences of bioclim variables (named with the variables)
}
\description{
Get bioclim seq
}
