% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coef.R
\name{get_coef}
\alias{get_coef}
\title{Get coefficients}
\usage{
get_coef(basename, sp_list, digits = 5)
}
\arguments{
\item{basename}{File basename (part before species name)}

\item{sp_list}{List of species directories to read occupancy from}

\item{digits}{integer indicating the number of decimal places to be kept.}
}
\value{
A \code{data.frame} with the grid_id in rows and country in columns
}
\description{
Get coefficients
}
