% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_env.R
\name{get_env}
\alias{get_env}
\title{Merge the phenological data per day}
\usage{
get_env(dir_sp, digits = 5)
}
\arguments{
\item{dir_sp}{Directory with output of occupancy model (qs)}

\item{digits}{integer indicating the number of decimal places to be kept.}
}
\value{
A \code{data.frame} with the grid_id in rows and country in columns
}
\description{
Merge the phenological data per day
}
