% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_poly_occupancy.R
\name{get_poly_occupancy}
\alias{get_poly_occupancy}
\title{Attach the occupancy outputs to the grid}
\usage{
get_poly_occupancy(grid, sp_list, digits = 5, verbose = TRUE)
}
\arguments{
\item{grid}{a \code{terra::SpatVector} object with the grid}

\item{sp_list}{List of species directories to read data from}

\item{digits}{integer indicating the number of decimal places to be kept.}

\item{verbose}{Print messages?}
}
\value{
A \code{terra::SpatVector} with occupancy summarized per species
}
\description{
Attach the occupancy outputs to the grid
}
