% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_slope}
\alias{get_slope}
\title{Get slope}
\usage{
get_slope(x, y, varx = NULL)
}
\arguments{
\item{x}{x: predictor}

\item{y}{y: variable to predict}

\item{varx}{optional precomputed variance of x}
}
\value{
exp(a) - 1 where b is the least-squares estimate of the regression slope
in the model y = a*x + b
}
\description{
Get exp(a) - 1, where a is the slope estimator of the linear model y ~ x
}
