% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ts_country.R
\name{get_ts_country}
\alias{get_ts_country}
\title{Calculate the average occupancy per country and per time step}
\usage{
get_ts_country(grid, sp_list, country = dragon_country(), digits = 5)
}
\arguments{
\item{grid}{a \code{terra::SpatVector} object with the grid}

\item{sp_list}{List of species directories to read occupancy from}

\item{country}{a \code{terra::SpatVector} object with the country definition}

\item{digits}{integer indicating the number of decimal places to be kept.}
}
\value{
A \code{data.frame} with the grid_id in rows and country in columns
}
\description{
Calculate the average occupancy per country and per time step
}
