% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{last_str}
\alias{last_str}
\title{get the last element of a string}
\usage{
last_str(x, split = ";")
}
\arguments{
\item{x}{string vector}

\item{split}{separator}
}
\description{
get the last element of a string
}
