% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_base_map}
\alias{plot_base_map}
\title{Plot base map}
\usage{
plot_base_map(map_data, Zmin, Zmax, bounds)
}
\arguments{
\item{map_data}{the map data}

\item{Zmin}{min zoom}

\item{Zmax}{max zoom}

\item{bounds}{bounds (xmin, ymin, xmax, ymax)}
}
\value{
a leaflet map
}
\description{
Plot base map
}
