% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_ly_lines}
\alias{plot_ly_lines}
\title{plot_ly_lines}
\usage{
plot_ly_lines(df)
}
\arguments{
\item{df}{dataframe with plotting data. Must have columns x, qmin, qmax, median and popup.}
}
\value{
A plot_ly plot with lines and error ribbon
}
\description{
plot_ly_lines
}
