% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_ly_scatter}
\alias{plot_ly_scatter}
\title{plot_ly_scatter}
\usage{
plot_ly_scatter(df)
}
\arguments{
\item{df}{dataframe with plotting data. Must have columns x, qmin, qmax, dmin, dmax, median and popup.}
}
\value{
A plot_ly plot with pojnts and error bars
}
\description{
plot_ly_scatter
}
