% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_polygons_map}
\alias{plot_polygons_map}
\title{Plot polygons on map}
\usage{
plot_polygons_map(
  leaflet_map,
  map_type,
  map_data,
  leg_names,
  layer_id = "mapid"
)
}
\arguments{
\item{leaflet_map}{A leaflet map (e.g. created with \code{plot_base_map})}

\item{map_type}{Map type (slope or average)}

\item{map_data}{Data to use for plotting polygons}

\item{leg_names}{A named vector for the legend names
(names are the titles to display, values correspond to \code{map_type})}

\item{layer_id}{layer id for the plotted polygons}
}
\value{
A leaflet map
}
\description{
Plot polygons on map
}
