% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_trend}
\alias{plot_trend}
\title{Plot trend}
\usage{
plot_trend(dat_psi)
}
\arguments{
\item{dat_psi}{A data.frame with columns year, mean, qmin, qmax, country}
}
\value{
A ggplot (created with ggiraph) with one trend line + CI pre country
}
\description{
Plot occupancy trends
}
