% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_shiny_data.R
\name{add_shiny_data}
\alias{add_shiny_data}
\title{Calculate the average occupancy per country and per time step}
\usage{
add_shiny_data(
  dir_sp,
  grid_file,
  country = dragon_country(),
  overwrite = FALSE
)
}
\arguments{
\item{dir_sp}{Directory with output of occupancy model (qs)}

\item{grid_file}{Path to a grid file (gpkg object)}

\item{country}{a \code{terra::SpatVector} object with the country definition}

\item{overwrite}{whether existing data will be overwritten}
}
\value{
A \code{data.frame} with the grid_id in rows and country in columns
}
\description{
Calculate the average occupancy per country and per time step
}
