% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{addLegend_decreasing}
\alias{addLegend_decreasing}
\title{Customized leaflet::addLegend() function to add a color legend to a map}
\usage{
addLegend_decreasing(
  map,
  position = c("topright", "bottomright", "bottomleft", "topleft"),
  pal,
  values,
  na.label = "NA",
  bins = 7,
  colors,
  opacity = 0.5,
  labels = NULL,
  labFormat = leaflet::labelFormat(),
  title = NULL,
  className = "info legend",
  layerId = NULL,
  group = NULL,
  data = leaflet::getMapData(map),
  decreasing = FALSE
)
}
\arguments{
\item{map}{a leaflet map}

\item{position}{the position of the legend}

\item{pal}{the color palette function, generated from
\code{\link[leaflet:colorNumeric]{leaflet::colorNumeric()}}, \code{colorBin()}, \code{colorQuantile()}, or
\code{colorFactor()}}

\item{values}{the values used to generate colors from the palette function}

\item{na.label}{the legend label for \code{NA}s in \code{values}}

\item{bins}{an approximate number of tick-marks on the color gradient for the
\code{colorNumeric} palette if it is of length one; you can also provide a
numeric vector as the pre-defined breaks (equally spaced)}

\item{colors}{a vector of (HTML) colors to be used in the legend if
\code{pal} is not provided}

\item{opacity}{the opacity of colors}

\item{labels}{a vector of text labels in the legend corresponding to
\code{colors}}

\item{labFormat}{a function to format the labels derived from \code{pal} and
\code{values} (see Details below to know what \code{labelFormat()} returns
by default; you can either use the helper function \code{labelFormat()}, or
write your own function)}

\item{title}{the legend title}

\item{className}{extra CSS classes to append to the control, space separated}

\item{layerId}{the ID of the legend; subsequent calls to \code{addLegend()}
or \code{addControl()} with the same \code{layerId} will replace this
legend. The ID can also be used with \code{removeControl()}.}

\item{group}{\code{group} name of a leaflet layer group.
Supplying this value will tie the legend to the leaflet layer group
with this name and will auto add/remove the legend as the
group is added/removed, for example via \code{layerControl()}.
You will need to set the \code{group} when you add a layer
(e.g., \code{\link[leaflet:map-layers]{leaflet::addPolygons()}}) and supply the same name here.}

\item{data}{data from the map}

\item{decreasing}{whether the legend should be in decreasing order}
}
\description{
When a color palette function is used in a map (e.g.,
\code{\link[leaflet:colorNumeric]{leaflet::colorNumeric()}}), a color legend can be automatically derived from
the palette function. You can also manually specify the colors and labels for
the legend.
}
\details{
The \code{labFormat} argument is a function that takes the argument
\code{type = c("numeric", "bin", "quantile", "factor")}, plus, arguments for
different types of color palettes. For the \code{colorNumeric()} palette,
\code{labFormat} takes a single argument, which is the breaks of the numeric
vector, and returns a character vector of the same length. For
\code{colorBin()}, \code{labFormat} also takes a vector of breaks of length
\code{n} but should return a character vector of length \code{n - 1}, with
the \code{i}-th element representing the interval \code{c(x[i], x[i + 1])}.
For \code{colorQuantile()}, \code{labFormat} takes two arguments, the quantiles
and the associated probabilities (each of length \code{n}), and should return
a character vector of length \code{n - 1} (similar to the \code{colorBin()}
palette). For \code{colorFactor()}, \code{labFormat} takes one argument, the
unique values of the factor, and should return a character vector of the same
length.

By default, \code{labFormat} is basically \code{format(scientific = FALSE, big.mark = ",")} for the numeric palette, \code{as.character()} for the
factor palette, and a function to return labels of the form \samp{x[i] - x[i
+ 1]} for bin and quantile palettes (in the case of quantile palettes,
\code{x} is the probabilities instead of the values of breaks).
}
