% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragonapp.occupancy-package.R
\docType{package}
\name{dragonapp.occupancy-package}
\alias{dragonapp.occupancy}
\alias{dragonapp.occupancy-package}
\title{dragonapp.occupancy: Research compendium to visually explore spatio-temporal occupancy model outputs}
\description{
This research compendium creates a shiny app and imports occupancy outputs. The Shiny app visually explore the spatial and temporal dynamics of the species distribution model.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Dragon-odonates/dragonapp-occupancy}
  \item Report bugs at \url{https://github.com/Dragon-odonates/dragonapp-occupancy/issues}
}

}
\author{
\strong{Maintainer}: Romain Frelat \email{rom.frelat@gmail.com} (\href{https://orcid.org/0000-0002-8631-4398}{ORCID})

Authors:
\itemize{
  \item Lisa Nicvert \email{lisa.nicvert@fondationbiodiversite.fr} (\href{https://orcid.org/0009-0006-5763-0865}{ORCID})
}

}
\keyword{internal}
