% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_poly_occupancy.R
\name{get_poly_occupancy}
\alias{get_poly_occupancy}
\title{Attach the occupancy outputs to the grid}
\usage{
get_poly_occupancy(grid, dir_sp, digits = 5)
}
\arguments{
\item{grid}{a \code{terra::SpatVector} object with the grid}

\item{dir_sp}{directory with output of occupancy model (qs)}

\item{digits}{integer indicating the number of decimal places to be kept.}
}
\value{
A \code{terra::SpatVector} with occupancy summarized per species
}
\description{
Attach the occupancy outputs to the grid
}
